// data/github/current/ui/packages/memex/src/client/api/memex-items/paginated-views.ts
export type PaginatedItemsData = BasePaginatedData & {
  nodes: Array<MemexItem>
  pageInfo: PageInfo
}

export type PaginatedGroupsData = BasePaginatedData & {
  groups: PaginatedGroups
  groupedItems: Array<GroupedMemexItems<MemexItem>>
  secondaryGroups: null
}

export type PaginatedGroupsAndSecondaryGroupsData = Omit<PaginatedGroupsData, 'secondaryGroups'> & {
  secondaryGroups: PaginatedGroups
}

export type PaginatedMemexItemsData = PaginatedItemsData | PaginatedGroupsData | PaginatedGroupsAndSecondaryGroupsData

type BasePaginationParameters = {
  /** Server-generated cursor indicating where we want to start the items / groups returned */
  after?: string
  /** Maximum number of items / groups to return */
  first?: number
}

type SecondaryGroupPaginationParameters = {
  /** Server-generated cursor indicating where we want to start the secondary groups returned.
   * Only applies if the current view is grouped in two directions (e.g. swimlanes in a board view)
   */
  secondaryAfter?: string
}

type ItemPaginationParameters = BasePaginationParameters & {
  /** Server-generated cursor indicating where we want to start the items / groups returned, working backwards */
  before?: string
  /** Maximum number of items / groups to return */
  last?: number
}

type SliceByParameters = {
  /** Field by which the view is sliced */
  sliceByColumnId?: MemexProjectColumnId
  /** Value for which we want to filter items */
  sliceByValue?: string
}

type SharedViewRequestParameters = {
  /** Query string from the filter for a view */
  q?: string
  /** Whether to return all, archived, or unarchived items */
  scope?: PaginatedItemsScope
  /** Field and direction by which the view is sorted */
  sortedBy?: Array<{columnId: number; direction: 'asc' | 'desc'}>
  /** A subset of field ids for the view, based on the visible fields */
  fieldIds?: ReadonlyArray<SystemColumnId | number>
  /** The type (i.e., layout) of the view */
  viewType?: ViewType
  /** A list of field ids to return sum aggregations for */
  sumFields?: Array<MemexProjectColumnId>
}

type HorizontalGroupByParameters = {
  /** Field used for horizontal grouping in the view */
  horizontalGroupedByColumnId?: MemexProjectColumnId
  /** Horizontal group value for which we want to request more items */
  groupedByGroupValue?: string
}

type VerticalGroupByParameters = {
  /** Field used for vertical grouping in the view */
  verticalGroupedByColumnId?: MemexProjectColumnId
  /** Vertical group value for which we want to request more items */
  verticalGroupedByGroupValue?: string
}
// ...
export type GetPaginatedItemsRequest = SharedViewRequestParameters &
  SliceByParameters &
  ItemPaginationParameters &
  SecondaryGroupPaginationParameters &
  HorizontalGroupByParameters &
  VerticalGroupByParameters

/**
 * Shape of response to a request to the `paginated_items` endpoint.
 * Can contain items, groups, secondary groups, grouped items, slice
 * by data, pagination information as well as total count data, depending
 * on the parameters in the request.
 */
export type GetPaginatedItemsResponse = PaginatedMemexItemsData